Add SourceLocation struct
Rather than manually add a row and col field to every individual class
that needs it, instead propogate a single SourceLocation struct that
contains both. We can just use one class since it's likely a row or col
will never be used on their own. We also have one source of truth where
we control the type of a row and col.

At some point, we'll also add a SourceLocation representing the end
position of a token.